requirements.txt
flask~=2.1.2
flask_login
flask_migrate
WTForms~=3.0.1
flask_wtf
flask_sqlalchemy
sqlalchemy
email_validator
python-decouple
gunicorn
jinja2~=3.1.2
flask-restx
Werkzeug~=2.1.2

pandas~=1.4.2
numpy~=1.22.4
nltk~=3.7
zeyrek~=0.1.2
snowballstemmer~=2.2.0
tensorflow~=2.9.1